Improve accessability of UI
Because vmsg appends its controls to the end of the DOM, it is not easy to
focus to the record button when the dialog is launched as keyboard-users 
need to tab to through the page to the end. Adding `recordBtn.focus()` jumps
the focus point right to the dialog, aiding navigation when then dialog is opened.

Additionally, when recording is started, hiding the record button can cause focus
to jump to the beginning of the page, re-focusing on the stop button ensures that
keyboard users have the stop button focused without re-navigating to it. Similarly,
when stopping recording, shift the focus back to the newly visible record button.

In addition to these focus changes, add title text to the buttons. Their purpose is
otherwise hidden from screen-readers.